Bulk issues fixed, needs integration
